requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
beautifulsoup4==4.8.2
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-autoslug==1.9.8
django-cogwheels==0.3
django-debug-toolbar==3.1.1
django-filter==2.4.0
django-modelcluster==5.1
django-recaptcha==2.0.6
django-taggit==1.3.0
django-treebeard==4.3.1
djangorestframework==3.12.1
draftjs-exporter==2.1.7
et-xmlfile==1.0.1
future==0.18.2
html5lib==1.1
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
l18n==2018.5
lml==0.0.9
oauth2client==4.1.3
odfpy==1.3.5
openpyxl==3.0.5
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyexcel-io==0.6.2
pyexcel-ods==0.5.3
PyPDF2==1.26.0
python-bidi==0.4.2
pytz==2020.1
reportlab==3.5.53
requests==2.24.0
rsa==4.6
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tablib==2.0.0
Unidecode==1.1.1
urllib3==1.25.10
wagalytics==1.2
wagtail==2.10.2
wagtail-metadata==3.3.0
wagtailfontawesome==1.2.1
wagtailmenus==3.0.2
wagtailstreamforms==3.11
webencodings==0.5.1
Willow==1.4
xhtml2pdf==0.2.5
xlrd==1.2.0
XlsxWriter==1.3.6
xlwt==1.3.0
